# Micro optimization: `slice(0...-3)` is actually faster than `slice!(-3, 3)`