@fxn 2020-01-27
Good one! 

I believe we could go with `byteslice` then, since the `ruby?` predicate is

```ruby
def ruby?(path)
  path.end_with?(".rb")
end
```

and guards that code.