@casperisfine 2020-01-27
> I believe we could go with `byteslice`

Yes, it just seemed a bit weird, so I wasn't sure. I'll update the PR.