@casperisfine 2020-01-27
BTW for posterity's sake, I got a fix for the extra `slice!` allocation: https://github.com/Shopify/ruby/commit/12838feadcfb4f394f3a93134cfbe76fb660f375

TL;DR; `slice!(range)` internally calls `self[range] = ''` which causes that extra allocation we see.

I'll see about submitting it upstream, but I'm really not sure such small perf fix would be accepted.