@casperisfine 2020-01-27
Damn, wait a bit. While writing this I just realized that `basename[]= ''` could actually be faster. Let me update my benchmark.