@casperisfine 2020-01-27
So on the micro benchmark it's very sligthly slower than byteslice
```
       slice!(-3, 3)      4.915M (± 1.3%) i/s -     24.821M in   5.050738s
   byteslice(0...-3)      8.125M (± 0.6%) i/s -     40.815M in   5.023349s
       [-3..-1] = ""      7.842M (± 0.7%) i/s -     39.249M in   5.005565s
      delete_suffix!     10.920M (± 0.5%) i/s -     54.864M in   5.024335s
```

However it has the advantage of being entirely allocation free:
```
>> s = "foo.rb"; MemoryProfiler.report { s[-3..-1] = ''.freeze }.pretty_print
Total allocated: 0 bytes (0 objects)
Total retained:  0 bytes (0 objects)
```

Up to you if you'd like to use that last alternative.