@casperisfine 2020-01-27
Welcome. TBH it was more about investigating that weird allocation than actually making anything noticeably faster.